LeapDesktopApplicationFull - V2.0.7
Switching All images to be loaded from resource file and based on that
the location from global location instead of local to my machine.

Change function defination of createUserFeedBackWindow from string
cinder::DataSourceRef instead to make use of the resources. This
included changing many of the other places where the function was
called.